Bug #1640199 bsp: camera: Coverity code optimization
[root cause  ] 1333583;1333574;1333573;1333006;1332164
[changes     ] isp2.6
[side effects] na
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:l6/n6pro
[test case]:na
[reviewers   ] eddy.wei
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

+++ core/isp2.6/src/isp_3dnr.c
@@ -446,8 +446,8 @@ static int isp3dnr_fbd_fetch_config_gen(struct isp_3dnr_ctx_desc *ctx)
 		dcam_if_cal_compressed_addr(&cal_fbc);
 		fbd_fetch->frame_header_base_addr = fbd_fetch->hw_addr.addr0;
 		fbd_fetch->slice_start_header_addr = fbd_fetch->frame_header_base_addr +
-			((fbd_fetch->slice_start_pxl_ypt / ISP_FBD_TILE_HEIGHT) * fbd_fetch->tile_num_pitch +
-			fbd_fetch->slice_start_pxl_xpt / ISP_FBD_TILE_WIDTH) * 16;
+			(((unsigned long)fbd_fetch->slice_start_pxl_ypt / ISP_FBD_TILE_HEIGHT) * fbd_fetch->tile_num_pitch +
+			(unsigned long)fbd_fetch->slice_start_pxl_xpt / ISP_FBD_TILE_WIDTH) * 16;
 	} else {
 		isp_3dnr_cal_compressed_addr(cur_width, cur_height,
 			ctx->buf_info[1]->iova[0], &out_addr);
@@ -461,8 +461,8 @@ static int isp3dnr_fbd_fetch_config_gen(struct isp_3dnr_ctx_desc *ctx)
 		dcam_if_cal_compressed_addr(&cal_fbc);
 		fbd_fetch->frame_header_base_addr = fbd_fetch->hw_addr.addr0;
 		fbd_fetch->slice_start_header_addr = fbd_fetch->frame_header_base_addr +
-			((fbd_fetch->slice_start_pxl_ypt / ISP_FBD_TILE_HEIGHT) * fbd_fetch->tile_num_pitch +
-			fbd_fetch->slice_start_pxl_xpt / ISP_FBD_TILE_WIDTH) * 16;
+			(((unsigned long)fbd_fetch->slice_start_pxl_ypt / ISP_FBD_TILE_HEIGHT) * fbd_fetch->tile_num_pitch +
+			(unsigned long)fbd_fetch->slice_start_pxl_xpt / ISP_FBD_TILE_WIDTH) * 16;
 	}
 
 	fbd_fetch->y_tiles_num_pitch = pad_width / FBD_NR3_Y_WIDTH;
